changed double to bigdecimal

+++ store/src/main/java/practice/store/exceptions/product/ProductFinalPriceException.java
@@ -1,10 +1,12 @@
 package practice.store.exceptions.product;
 
+import java.math.BigDecimal;
+
 public class ProductFinalPriceException extends RuntimeException {
 
-    public ProductFinalPriceException(double finalPrice, double finalPriceCalculate) {
+    public ProductFinalPriceException(BigDecimal finalPrice, BigDecimal finalPriceCalculate) {
         super(String.format(
-                "Incorrect final price. Final price from payload:%f. Correct final price:%f.",
+                "Incorrect final price. Final price from payload:%.2f. Correct final price:%.2f.",
                 finalPrice,
                 finalPriceCalculate
         ));
